server/requirements.txt
appdirs==1.4.4
audioread==2.1.9
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
decorator==5.1.1
Flask==2.1.2
Flask-Cors==3.0.10
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kafka-python==2.0.2
librosa==0.9.2
llvmlite==0.38.1
MarkupSafe==2.1.1
numba==0.55.2
numpy==1.22.4
packaging==21.3
pooch==1.6.0
pycparser==2.21
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.1
resampy==0.3.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
SoundFile==0.10.3.post1
threadpoolctl==3.1.0
urllib3==1.26.10
Werkzeug==2.1.2
